Maint: Gemfile/.gitignore
This adds a .gitignore and a Gemfile to the repository. The gemfile is based on
the usage of puppet/facter/hiera from a local directory if existing (through
environment variables). Without this a developer may have to install
puppet/facter prior to working on this and then uninstall them after to work on
those repositories.