(#19162) Gracefully handle when shutdown.exe command cannot be found
Previously, we were executing the x86 version of the shutdown command on x64
Windows. On non-R2 Windows 2003 x64 systems, there is no x86 version and the
sysnative alias is not available without hotfix 942589 installed.

This commit changes the provider to prefer the x64 version of shutdown.exe on
x64 versions of Windows. If the sysnative alias is not available (x86 OS or
2003 x64 without the hotfix), then we fall back to the default `system32`
location. On 2003 non-R2 x64, that will not exist either, so we raise an
exeption.

Paired-with: Rob Reynolds <rob@puppetlabs.com>