(#19162) Restrict example group to systems with the shutdown command
Previously, the `when a reboot is triggered` example group would fail on
Windows 2003 (non-R2) x64, because we could not resolve shutdown.exe into a
fully qualified path, and we would correctly raise an exception.

This commit only executes those tests on systems where the shutdown command can
be resolved.

Paired-with: Rob Reynolds <rob@puppetlabs.com>