(PE-1669) Never load sample.pp in production
Somehow sample.pp was being loaded by the parser, and then included in
catalogs for *nix agents, because the resources are not contained in a
class. Exactly how they were getting included is a mystery.

However, since the manifest is just used to test the reboot module, this
commit moves the manifest to the tests directory.