(MODULES-447) Retry shutdown abort command
Windows does not have an API for determining if a shutdown is currently in
progress. Instead you can try to abort any in progress shutdown, and if the
command (shutdown /a) returns 0, then it means a shutdown was successfully
aborted, or it will return 1116, which means no shutdown was in progress.

Note due to a ruby bug[1], it truncates exit code 1116 to a single byte,
resulting in 92.

However, puppet executes the watcher process asynchronously, so it was
possible for the acceptance test to issue the abort shutdown command
before the watcher process had initiated the shutdown. The tests tried
to work around this by sleeping for 5 seconds, but this doesn't work
on slow systems in CI.

This commit adds a helper method for retrying the shutdown abort command
and updates the various acceptance tests to use it. It also adds a
helper method for the opposite case, when verifying that a shutdown has
NOT been initiated.

[1] https://bugs.ruby-lang.org/issues/8083