(MODULES-447) Handle race when puppet has already exited
Previously, if the watcher called OpenProcess and it failed, we couldn't open
a handle to the puppet process, and would treat that as a fatal error.

However, if puppet exits before the watcher process called OpenProcess, then
the watcher would never reboot the system.

This commit checks to see if OpenProcess fails and GetLastError returns 87
(INVALID_PARAMTER). If so, then it means the pid no longer exists. And we can
conclude that puppet has already exited.

However, if OpenProcess fails for some other reason, e.g. access denied, then
we continue to treat that as a fatal error.