(MODULES-447) Don't write to stderr from the watcher
Previously, the watcher was writing the shutdown command to $stderr just
before executing the command. However, if the puppet process had already
exited, then the call to $stderr.puts will raise an exception when
executed via cygwin, and the shutdown would never happen.

The exception message reported: Invalid argument - <STDERR>. Also the
exception wasn't being caught, instead the watcher would silently exit.

Note this issue is timing dependent, since everything works out ok if puppet
is still running when the watcher writes the shutdown command to $stderr.

This commit adds exception handling around the watcher, and changes the
watcher to write to a temp file. Also some of the acceptance tests rely
on parsing the shutdown command that the watcher is going to execute. I
moved this into the provider itself.

One acceptance test had to be modified so that it doesn't try to match
the timeout message, since the watcher no longer writes to $stderr.