(MODULES-1729) Round :timeout up to nearest minute
Linux uses minutes to specify its shutdown timeout, but the :timeout
parameter is in seconds. Ruby uses integer division, so previously
values of :timeout that weren't whole multiples of 60 would be rounded
down. This changes the behavior so that they are rounded up and a
warning is issued notifying the user.