(MODULES-2804) Fix use of Win32::Registry.read
The `read` method of `Win32::Registry` class returns a `[type,
data]` array if the value exists, but it's used as if it returns just the
`data` in `package_installer?` and `pending_computer_rename`.

In most cases, there is no problem with it because the detecting result is
determined by whether the registry value exists or not, not by the `data`.
However it may cause issues sometimes. Instead be sure to explicitly grab
the data from the return value when necesary.