(maint) Update rakefile for puppetlabs_spec_helper
The rakefile is attempting to override the default puppet lint settings however
these settings are more permissive than the defaults and are causing false
lint failures e.g. the vendor directory is not excluded in thie Rakefile but
the puppetlabs_spec_helper does this by default.  This commit removes the
ignore_paths setting as the spec helper already does this.  Also added a comment
that the 80 and 140 char rule changes can be removed once next version of
spec_helper is released as they are already the defaults.