(maint) Fix Gemfile for PE 2016.2
The gem specification for mimetypes was causing an old version of beaker to be installed.
This version of beaker is not compatible with 2016.2 as it is unable to use the beaker-answers
gem as 2016.2 is installed using MEEP.  The mimetypes specification appears to be a hold over
from copying sync.yml from the acl project.  This commit keeps the nokogiri gem requirement
but uses the optional gems notation in the sync.yml so that the defaults from modsync still
get applied.

This commit also updates Gemfile with modsync.