(maint) Return timeout in seconds
Previously the reboot task would report the timeout in seconds for
SunOS, and minutes for any other posix system. Change the task to report
seconds for all systems, with non-SunOS systems adjusted to the actual
restart time (as the shutdown delay on those systems can only be
expressed in minutes and hours).