(maint) Wait requested time on Linux, return timeout in seconds
Update the timeout task to wait the requested number of seconds on Linux
rather than rounding up to the nearest minute. Also switch from
reporting the timeout in minutes for Linux to seconds, so its consistent
with SunOS and Windows.

Also the task previously wouldn't run over PE on Linux systems because
it had no shebang, resulting in exit code 8. Add the shebang to use
Puppet Agent's Ruby.