(packaging) Make task implementations unique for the Forge and PE
Older versions of PE won't accept tasks with multiple implementations
that share the same script basename. The Forge enforces this as a rule
for modules uploaded to it, and also requires that every task have
metadata. Make the implementations use different names and have their
own metadata as well as the original combined task.