(maint) Handle an empty $nodes argument cleanly
Prior to this commit the reboot plan raised an error calling max()
without a valid argument. This commit causes the plan to short-circuit
and return an empty ResultSet in the event an empty TargetSpec was given
as the $nodes argument.

Long-term, the plan should probably be modified to always return a
meaningful ResultSet value on completion.