(MODULES-9228) Return a ResultSet from reboot plan
Assuming some nodes successfully reboot and some nodes don't, when the
reboot plan is invoked with _catch_errors => true, it should return
meaningful data so that the end user can decide how to handle or report
on which nodes failed to reboot while potentially continuing forward
to another task with the set of nodes that succeeded.

This commit makes the reboot plan always return meaningful data, even
when erroring.

Side note: gonna open a ticket to make it easier to _catch_errors for a
plan, for the end user...