(MODULES-9228) Add fail_plan_on_errors flag
A common use case for using the reboot plan with _catch_errors => true
would be to review which systems didn't reboot and remediate or move
forwards only with those that did. However, even with _catch_errors =>
true, the caller has to check if the plan result is an error and go down
a different code path if so to deal with it and extract the ResultSet
they need.

This commit provides a new Boolean plan option, fail_plan_on_error,
which if set to false will ensure the plan always returns exactly the
same data type regardless of whether or not any individual target failed
to reboot. The consuming plan author can then simplify their own code,
secure in the knowledge of what kind of data will be returned.