Fix missed fail_plan() and provide legacy return
This commit fixes a missed fail_plan() invocation that would have
allowed a plan to fail even if fail_plan_on_errors was set to false. In
doing so, it further linearizes the logic, allowing for a slight
simplification of the main reduce function.

While making the commit, it seemed like a good idea to restore the
fail_plan() detail key that was used previously, in order to retain
compatibility with previous failure return results.

Clean up the spec test a bit while in here.