(IAC-140) Refactor Spec Helper code for new Litmus exec'd tests
There was a significant refactor in the test behaviour for Linux
systems as part of the port to Litmus. Windows remains largely
unchanged.

For Linux systems, we significantly increased the platforms we're
testing against to more accurately reflect the module's compatible
OSs. Relying on 'shutdown -c' was not possible anymore:
- systemd based OSs create /run/nologon when shutdown is invoked,
preventing any more logins during the test
- On modern OSs, 'shutdown -c' cannot cancel the reboot operation
in anywhere near quick enough a manner
- 'shutdown -c' will return 0 even if a shutdown operation was NOT
in progress
- 'shutdown -c' did not have consistent stdout / stderr messages
to validate against. Typically, shutdown is a symlink to another
binary / script, so behaviour differed between OSs quite sigificantly.

The new approach is to substitute the shutdown cmd on PATH with a
script that will output the args it was called with to a log file.
We will then verify the correct args were passed to the script.
This allows us to execute the agent with the module on each platform
and ensure that functionality such as subscribing to another resource
or invoking another resource downstream of a reboot, still works as
expected.

We will take a backup of the shutdown bin / script and restore it
after the test suite has executed. There is also logic to ensure
that we do not overwrite this binary if our dummy script is still
present (this handles the scenario where a test is interrupted and
then resumed before the RSpec.after block has executed).