(MODULES-10955) More robust handling of reboot-task output
The reboot task (that gets run in line 30) always returns the status and timeout. The timeout is (per definition) the max between 3 and $reboot_delay.
No real dynamic output that is generated by the reboot task is used in the plan.

Hence: We can default to a $wait for the maximum of 3 or $reboot_delay. This way it does not matter to the plan what the reboot task runs.
So broadcast-messages (and possibly other things) do not interfere with the parsing of the timeout.